# api/openapi.yaml
openapi: "3.1.0"
info:
  title: OpenClause API
  description: >
    Single entrypoint for AI agent tool-call governance.
    Validates requests, evaluates policy, manages approvals, and records audit evidence.
  version: "1.0.0"
  contact:
    name: OpenClause Team
servers:
  - url: http://localhost:8080
    description: Local gateway
  - url: http://localhost:8081
    description: Local approvals service
  - url: http://localhost:8090
    description: Local console API
security:
  - ApiKeyAuth: []
paths:
...
  /v1/integrations/slack/interactions:
    post:
      operationId: slackInteractions
      summary: Receive Slack interactive approval actions
      tags: [Approvals]
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                payload:
                  type: string
      responses:
        "200":
          description: Slack message update payload
        "401":
          description: Invalid Slack signature
        "403":
          description: Approver not allowed for tenant

  # ── Console Onboarding ──────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...